Refactored entire code base made it modular and more flexible
